Rename AppointmentBlockWithIcons to AppointmentBlock
