+++ src/applications/check-in/components/AppointmentBlock.jsx
